requirements.txt
streamlit==1.32.0
langgraph==0.1.19
langchain==0.1.20
langchain-core==0.1.52
openai==1.30.1
pdfplumber==0.11.0
python-docx==1.1.0
fpdf2==2.7.9
pydantic==2.7.1
python-dotenv==1.0.1